+++ plugins/tavern-core/src/main/kotlin/com/loyea/plugins/tavern/core/WorldInfoModels.kt
@@ -1,4 +1,4 @@
-package com.loyea.ui.chat
+package com.loyea.plugins.tavern.core
 
 enum class WorldInfoInsertionOrder {
     ORDER,