+++ plugins/tavern-core/src/test/kotlin/com/loyea/plugins/tavern/core/TavernPluginRuntimeTest.kt
@@ -1,4 +1,4 @@
-package com.loyea.ui.chat
+package com.loyea.plugins.tavern.core
 
 import com.loyea.plugin.api.ChatRole
 import com.loyea.plugin.api.ConversationText